FEATURE: parsing parse light; parse sphere; parse basics including color, brightness and position; parse utils contains is_all_numbers and validations for color range and argument count
