+++ src/parse_lights.c
@@ -6,105 +6,25 @@
 /*   By: krocha-h <krocha-h@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/13 20:15:43 by krocha-h          #+#    #+#             */
-/*   Updated: 2024/10/15 20:37:28 by krocha-h         ###   ########.fr       */
+/*   Updated: 2024/10/19 10:06:09 by krocha-h         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-bool	is_all_numbers(char **split)
-{
-	int	i;
-	int	j;
-
-	i = 0;
-	while (split[i])
-	{
-		j = 0;
-		if (split[i][j] == '-')
-			j++;
-		while (split[i][j] != '\0' && split[i][j] != '\n')
-		{
-			if (!ft_isdigit(split[i][j]) && split[i][j] != '.')
-				return (false);
-			j++;
-		}
-		i++;
-	}
-	return (true);
-}
-
-bool	validate_count(char **split, int count)
-{
-	int	i;
-
-	i = 0;
-	while (split[i])
-		i++;
-	if (i != count)
-		return (false);
-	return (true);
-}
-
-bool	validate_color_range(char **str)
-{
-	int	i;
-
-	i = 0;
-	while (str[i])
-	{
-		if (ft_atoi(str[i]) < 0 || ft_atoi(str[i]) > 255)
-			return (false);
-		i++;
-	}
-	return (true);
-}
-
-bool	parse_color(char *splitted, t_color *new_color)
-{
-	char	**colors;
-
-	*new_color = color(0, 0, 0);
-	colors = ft_split(splitted, ',');
-	if (!validate_count(colors, 3) || !is_all_numbers(colors)
-		|| !validate_color_range(colors))
-	{
-		ft_free_split(colors);
-		return (false);
-	}
-	*new_color = color(ft_atoi(colors[0]), \
-					ft_atoi(colors[1]), \
-					ft_atoi(colors[2]));
-	ft_free_split(colors);
-	return (true);
-}
-
-bool	parse_brightness(char *splitted, double *brightness)
-{
-	*brightness = ft_atof(splitted);
-	if (*brightness < 0 || *brightness > 1)
-		return (false);
-	return (true);
-}
-
-bool	parse_coordinates(char *splitted, t_point *position)
-{
-	char	**coordinates;
-
-	*position = point(0, 0, 0);
-	coordinates = ft_split(splitted, ',');
-	if (!validate_count(coordinates, 3) || !is_all_numbers(coordinates))
-		{
-			ft_free_split(coordinates);
-			return (false);
-		}
-	position->x = ft_atof(coordinates[0]);
-	position->y = ft_atof(coordinates[1]);
-	position->z = ft_atof(coordinates[2]);
-	ft_free_split(coordinates);
-	return (true);
-}
-
+/**
+ * @brief Parses a light definition from a line of input and adds it to the
+ * world.
+ *
+ * This function takes a line of input that defines a light and parses it to
+ * extract the light's properties. The parsed light is then added to the world.
+ *
+ * @param line A string containing the light definition.
+ * @param world A pointer to the world structure where the parsed light will be
+ * added.
+ * @return true if the light was successfully parsed and added to the world,
+ * false otherwise.
+ */
 bool	parse_light(char *line, t_world *world)
 {
 	t_color	new_color;
